requirements.txt
click
configparser
pandas
python-daemon
